+++ src/types_impl/lob/nclob.rs
@@ -139,7 +139,7 @@ impl NCLob {
 
     /// Current size of the internal buffer, in bytes.
     pub fn cur_buf_len(&self) -> usize {
-        self.0.cur_buf_len() as usize
+        self.0.cur_buf_len()
     }
 
     /// Provides information about the the server-side resource consumption that
@@ -252,7 +252,7 @@ impl NCLobHandle {
         }
 
         let read_length = std::cmp::min(
-            self.am_conn_core.lock()?.get_lob_read_length() as u32,
+            self.am_conn_core.lock()?.get_lob_read_length(),
             (self.total_char_length - self.acc_char_length as u64) as u32,
         );
